+++ src/ethereum/info/abstractInfo.ts
@@ -91,6 +91,8 @@ const currencyInfo: EdgeCurrencyInfo = {
     }
   ],
 
+  usesChangeServer: true,
+
   // Deprecated:
   defaultSettings: makeEvmDefaultSettings(networkInfo),
   displayName: 'Abstract',