+++ src/ethereum/info/bobevmInfo.ts
@@ -96,6 +96,8 @@ const currencyInfo: EdgeCurrencyInfo = {
     }
   ],
 
+  usesChangeServer: true,
+
   // Deprecated:
   defaultSettings: makeEvmDefaultSettings(networkInfo),
   displayName: 'BOB',