+++ src/ethereum/info/ethereumInfo.ts
@@ -1241,6 +1241,8 @@ export const currencyInfo: EdgeCurrencyInfo = {
     }
   ],
 
+  usesChangeServer: true,
+
   // Deprecated:
   defaultSettings: makeEvmDefaultSettings(networkInfo),
   displayName: 'Ethereum',