+++ src/ethereum/info/zksyncInfo.ts
@@ -139,6 +139,8 @@ const currencyInfo: EdgeCurrencyInfo = {
     }
   ],
 
+  usesChangeServer: true,
+
   // Deprecated:
   defaultSettings: makeEvmDefaultSettings(networkInfo),
   displayName: 'zkSync',